% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_lvl_input}
\alias{convert_lvl_input}
\title{Convert Log Level Input to Integer}
\usage{
convert_lvl_input(level)
}
\arguments{
\item{level}{Log level as character or numeric.}
}
\value{
The log level integer.
}
\description{
Convert Log Level Input to Integer
}
\keyword{internal}
