% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_csv}
\alias{convert_to_csv}
\title{Write log to csv file}
\usage{
convert_to_csv(file, logfile, remove_message_lf = TRUE, ...)
}
\arguments{
\item{file}{Path to write csv file to}

\item{logfile}{Path to log file to read from}

\item{remove_message_lf}{Should the line breaks at the end of messages be removed?}

\item{...}{Additional arguments to pass to \code{utils::write.table()}}
}
\value{
Invisible \code{NULL}.
}
\description{
Write log to csv file
}
