% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_call}
\alias{find_call}
\title{Find the Call of a Parent Function in the Call Hierarchy}
\usage{
find_call()
}
\value{
Returns the call of the parent function, or \code{NULL} if no such call is found.
}
\description{
This function is designed to inspect the call hierarchy and identify the call of a parent function.
Any wrapper environments above the global R environment that some IDEs cause are ignored.
}
\keyword{internal}
