% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_loc}
\alias{get_file_loc}
\title{Get file location}
\usage{
get_file_loc(x)
}
\arguments{
\item{x}{Call object.}
}
\value{
The file location as string.
}
\description{
Get the file location of a call object.
}
\keyword{internal}
