% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{get_logfile}
\alias{get_logfile}
\title{Get Log File}
\usage{
get_logfile()
}
\value{
The log file path.
}
\description{
Return the log file that \code{loggit()} will write to.
}
\examples{
get_logfile()

}
