% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lvl_int}
\alias{get_lvl_int}
\title{Get Log Level Integer}
\usage{
get_lvl_int(level)
}
\arguments{
\item{level}{Log level as character.}
}
\value{
The log level integer.
}
\description{
Get Log Level Integer
}
\keyword{internal}
