% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lvl_name}
\alias{get_lvl_name}
\title{Get Log Level Name}
\usage{
get_lvl_name(level)
}
\arguments{
\item{level}{Log level as integer.}
}
\value{
The log level name.
}
\description{
Get Log Level Name
}
\keyword{internal}
