% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_name}
\alias{get_package_name}
\title{Get package name}
\usage{
get_package_name(x)
}
\arguments{
\item{x}{Function.}
}
\value{
The package name as string.
}
\description{
Get the package name of a function.
}
\keyword{internal}
