% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggit.R
\name{loggit_internal}
\alias{loggit_internal}
\title{Internal logging function}
\usage{
loggit_internal(
  log_lvl,
  log_msg,
  log_call = NULL,
  echo,
  logfile = get_logfile(),
  call_options = get_call_options()
)
}
\arguments{
\item{log_lvl}{Log level. A atomic vector of length one (usually \code{character}). Will be coerced to \code{character}.}

\item{log_msg}{Log message. A atomic vector of length one (usually \code{character}). Will be coerced to \code{character}.}

\item{echo}{Should the log entry (json) be echoed to \code{stdout} as well?}

\item{logfile}{A \code{connection}, or a character string naming the file to write to.}
}
\value{
Invisible \code{NULL}.
}
\description{
This function is used internally by the \code{loggit} function to log messages and levels.
No checks are performed on the input, so it should used with caution.
}
\keyword{internal}
