% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_logs}
\alias{read_logs}
\title{Return log file as an R data frame}
\usage{
read_logs(logfile = get_logfile(), unsanitize = TRUE)
}
\arguments{
\item{logfile}{Path to log file.}

\item{unsanitize}{Should the log messages be unsanitized?}
}
\value{
A \code{data.frame}.
}
\description{
This function returns a \code{data.frame} containing all the logs in the provided \code{ndjson} log file.
}
\details{
\code{read_logs()} returns a \code{data.frame} with the empty character columns "timestamp", "log_lvl" and "log_msg"
if the log file has no entries.
}
\examples{
  set_logfile(file.path(tempdir(), "loggit.log"), confirm = FALSE)
  message("Test log message")
  read_logs()

}
