% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{read_ndjson}
\alias{read_ndjson}
\title{Read ndJSON-formatted log file}
\usage{
read_ndjson(logfile, unsanitize = TRUE, last_first = FALSE)
}
\arguments{
\item{logfile}{Path to log file to read from.}

\item{unsanitize}{Should escaped special characters be unescaped?}

\item{last_first}{Should the last log entry be the first row of the data frame?}
}
\value{
A \code{data.frame}, with the columns as the fields in the log file.
}
\description{
Read ndJSON-formatted log file
}
\keyword{internal}
