% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_logs}
\alias{rotate_logs}
\title{Rotate log file}
\usage{
rotate_logs(rotate_lines = 100000L, logfile = get_logfile())
}
\arguments{
\item{rotate_lines}{The number of log entries to keep in the logfile.}

\item{logfile}{Log file to truncate.}
}
\value{
Invisible \code{NULL}.
}
\description{
Truncates the log file to the line count provided as \code{rotate_lines}.
}
\examples{
\dontrun{
  rotate_logs()

  rotate_logs(rotate_lines = 0L)

  rotate_logs(rotate_lines = 1000L, logfile = "my_log.log")
}
}
