% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_call_options}
\alias{set_call_options}
\title{Set Call Options}
\usage{
set_call_options(..., .arg_list, confirm = TRUE)
}
\arguments{
\item{...}{Named arguments to set.}

\item{.arg_list}{A list of named arguments to set.}

\item{confirm}{Print confirmation message of call options setting?}
}
\value{
Invisible the previous call options.
}
\description{
Set Call Options
}
