% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_echo}
\alias{set_echo}
\title{Set echo}
\usage{
set_echo(echo = TRUE, confirm = TRUE)
}
\arguments{
\item{echo}{Logical. Should log messages be echoed to the console?}

\item{confirm}{Print confirmation message of echo setting?}
}
\value{
Invisible \code{NULL}.
}
\description{
Set echo
}
