% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_log_level}
\alias{set_log_level}
\title{Set Log Level}
\usage{
set_log_level(level = "INFO", confirm = TRUE)
}
\arguments{
\item{level}{Log level to set.}

\item{confirm}{Print confirmation message of log level?}
}
\value{
Invisible \code{NULL}.
}
\description{
Set Log Level
}
\details{
Log levels are as follows:
DEBUG: 4
INFO: 3
WARNING: 2
ERROR: 1
NONE: 0
}
