% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_timestamp_format}
\alias{set_timestamp_format}
\title{Set Timestamp Format}
\usage{
set_timestamp_format(ts_format = "\%Y-\%m-\%dT\%H:\%M:\%S\%z", confirm = TRUE)
}
\arguments{
\item{ts_format}{ISO date format.}

\item{confirm}{Print confirmation message of timestamp format?}
}
\value{
Invisible the previous timestamp format.
}
\description{
Set timestamp format for use in output logs.
}
\details{
This function performs no time format validations, but will echo out the current time in
the provided format for manual validation.

This function provides no means of setting a timezone, and instead relies on
the host system's time configuration to provide this. This is to enforce
consistency across software running on the host.
}
\examples{
\dontrun{
  set_timestamp_format("\%Y-\%m-\%d \%H:\%M:\%S")
}

}
