% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_call_options}
\alias{setup_call_options}
\title{Set up the call options for the application}
\usage{
setup_call_options()
}
\value{
Invisible the previous call options.
}
\description{
This function sets the call options for the application.
}
\keyword{internal}
