% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_echo}
\alias{setup_echo}
\title{Set up the echo for the application}
\usage{
setup_echo()
}
\value{
Invisible the previous echo setting.
}
\description{
This function retrieves the echo setting from the environment variable "ECHO_LOGGIT2".
If the environment variable is set, it will be used as the echo setting.
If the environment variable is not set, the echo setting will be set to the default value.
}
\details{
The environment variable must be set to one of: "TRUE" or "FALSE".
}
\keyword{internal}
