% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_log_level}
\alias{setup_log_level}
\title{Set up the log level for the application}
\usage{
setup_log_level()
}
\value{
Invisible the previous log level.
}
\description{
This function retrieves the log level from the environment variable "LEVEL_LOGGIT2".
If the environment variable is set, it will be used as the log level.
If the environment variable is not set, the log level will be set to the default value.
}
\details{
The environment variable must be set to one of: "DEBUG", "INFO", "WARN", "ERROR" or "NONE".
}
\keyword{internal}
