% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_timestamp_format}
\alias{setup_timestamp_format}
\title{Set up the timestamp format for the application}
\usage{
setup_timestamp_format()
}
\value{
Invisible the previous timestamp format.
}
\description{
This function retrieves the timestamp format from the environment variable "TIMESTAMP_LOGGIT2".
If the environment variable is set, it will be used as the timestamp format.
If the environment variable is not set, the timestamp format will be set to the default value.
}
\keyword{internal}
