% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_json.R
\name{split_ndjson}
\alias{split_ndjson}
\title{Convert a ndJSON object}
\usage{
split_ndjson(x)
}
\arguments{
\item{x}{A vector of JSON strings.}
}
\value{
A list with the same length as \code{x}, each containing two character vectors: \code{keys} and \code{values}.
}
\description{
Split a vector of JSON strings into a list of R objects.
}
\examples{
\dontrun{
  split_ndjson(c('{"a": "1", "b": "2"}', '{"a": "3", "b": "4", "c": "5"}'))
}
}
\keyword{internal}
