% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_loggit.R
\name{with_loggit}
\alias{with_loggit}
\title{Log any expressions}
\usage{
with_loggit(
  exp,
  logfile = get_logfile(),
  echo = TRUE,
  log_level = get_log_level()
)
}
\arguments{
\item{exp}{An expression to evaluate.}

\item{logfile}{Log file to write to. Defaults to currently-configured log
file.}

\item{echo}{Should the log file entry be printed to the console as well?
Defaults to \code{TRUE}.}

\item{log_level}{The log level to use.}
}
\value{
The result of the expression.
}
\description{
Log code without having to explicitly use the \code{loggit2} handlers.
This is particularly useful for code that cannot be customized, e.g. from third-party packages
}
\details{
If \code{loggit2} handlers are already used in the expression, this can lead to conditions being logged
twice (in the same or different files)
}
