% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_ndjson}
\alias{write_ndjson}
\title{Write ndJSON-formatted log file}
\usage{
write_ndjson(
  log_df,
  logfile = get_logfile(),
  echo = TRUE,
  overwrite = FALSE,
  sanitizer = default_ndjson_sanitizer
)
}
\arguments{
\item{log_df}{Data frame of log data. Rows are converted to \code{ndjson} entries,
with the columns as the fields.}

\item{logfile}{Log file to write to. Defaults to currently-configured log
file.}

\item{echo}{Echo the \code{ndjson} entry to the R console? Defaults to \code{TRUE}.}

\item{overwrite}{Overwrite previous log file data? Defaults to \code{FALSE}, and
so will append new log entries to the log file.}

\item{sanitizer}{\link[=sanitizers]{Sanitizer function} to run over elements in log data.
Defaults to \code{\link[=default_ndjson_sanitizer]{default_ndjson_sanitizer()}}.}
}
\description{
Write ndJSON-formatted log file
}
\keyword{internal}
