% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_2_string}
\alias{call_2_string}
\title{Convert Call to String}
\usage{
call_2_string(call, full_stack = FALSE)
}
\arguments{
\item{call}{Call object.}

\item{full_stack}{Include the full call stack?}
}
\value{
Deparsed call as string.
}
\description{
Converts a call object to a string and optionally determines the full call stack.
}
\details{
The full call stack can only be determined if the call is in the current context.
}
\keyword{internal}
