% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_echo}
\alias{set_echo}
\title{Set echo}
\usage{
set_echo(echo = TRUE, confirm = TRUE)
}
\arguments{
\item{echo}{Should log messages be echoed to \code{stdout}?}

\item{confirm}{Print confirmation message of echo setting?}
}
\value{
Invisible the previous echo setting.
}
\description{
Set echo
}
\examples{
\dontrun{
 set_echo(TRUE)
 set_echo(FALSE)
}

}
