% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{set_log_level}
\alias{set_log_level}
\title{Set Log Level}
\usage{
set_log_level(level = "DEBUG", confirm = TRUE)
}
\arguments{
\item{level}{Log level to set, as a string or integer.}

\item{confirm}{Print confirmation message of log level?}
}
\value{
Invisible the previous log level.
}
\description{
Set Log Level
}
\details{
Log levels are as follows:
DEBUG: 4
INFO: 3
WARNING: 2
ERROR: 1
NONE: 0
}
\examples{
\dontrun{
 set_log_level("DEBUG")
 set_log_level("INFO")

 set_log_level(4)
 set_log_level(3)
}

}
