% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_ndjson}
\alias{write_ndjson}
\title{Write ndJSON-formatted log file}
\usage{
write_ndjson(log_df, logfile = get_logfile(), echo = TRUE, overwrite = FALSE)
}
\arguments{
\item{log_df}{Data frame of log data. Rows are converted to \code{ndjson} entries,
with the columns as the fields.}

\item{logfile}{Log file to write to.}

\item{echo}{Echo the \code{ndjson} entry to the R console?}

\item{overwrite}{Overwrite previous log file data?}
}
\description{
Write ndJSON-formatted log file
}
\keyword{internal}
