% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_ndjson}
\alias{write_ndjson}
\title{Write ndJSON-formatted log file}
\usage{
write_ndjson(
  log_df,
  logfile = get_logfile(),
  echo = get_echo(),
  overwrite = FALSE
)
}
\arguments{
\item{log_df}{A \code{data.frame}. Entries are converted to \code{ndjson}, with the columns as the fields.}

\item{logfile}{Path of log file to write to.}

\item{echo}{Should the log entry (json) be echoed to \code{stdout} as well?}

\item{overwrite}{Overwrite previous log file?}
}
\value{
Invisible \code{NULL}.
}
\description{
Write ndJSON-formatted log file
}
\keyword{internal}
